Finalisation de la messagerie privée

+++ src/Controller/MessageController.php
@@ -3,9 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\Message;
+use App\Entity\User;
 use App\Form\MessageType;
 use App\Repository\MessageRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -17,23 +19,41 @@ final class MessageController extends AbstractController
     #[Route(name: 'app_message_index', methods: ['GET'])]
     public function index(MessageRepository $messageRepository): Response
     {
+        /** @var User $user */
+        $user = $this->getUser();
+
         return $this->render('message/index.html.twig', [
-            'messages' => $messageRepository->findAll(),
+            'messages' => $messageRepository->findBy(
+                ['receiver' => $user],
+                ['createdAt' => 'DESC']
+            ),
         ]);
     }
 
     #[Route('/new', name: 'app_message_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        /** @var User $user */
+        $user = $this->getUser();
+
         $message = new Message();
+
         $form = $this->createForm(MessageType::class, $message);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
+            $message->setSender($user);
+            $message->setCreatedAt(new \DateTimeImmutable());
+
             $entityManager->persist($message);
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_message_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute(
+                'app_message_index',
+                [],
+                Response::HTTP_SEE_OTHER
+            );
         }
 
         return $this->render('message/new.html.twig', [
@@ -45,21 +65,47 @@ final class MessageController extends AbstractController
     #[Route('/{id}', name: 'app_message_show', methods: ['GET'])]
     public function show(Message $message): Response
     {
+        /** @var User $user */
+        $user = $this->getUser();
+
+        if (
+            $message->getSender() !== $user &&
+            $message->getReceiver() !== $user
+        ) {
+            throw $this->createAccessDeniedException();
+        }
+
         return $this->render('message/show.html.twig', [
             'message' => $message,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_message_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Message $message, EntityManagerInterface $entityManager): Response
-    {
+    public function edit(
+        Request $request,
+        Message $message,
+        EntityManagerInterface $entityManager
+    ): Response {
+
+        /** @var User $user */
+        $user = $this->getUser();
+
+        if ($message->getSender() !== $user) {
+            throw $this->createAccessDeniedException();
+        }
+
         $form = $this->createForm(MessageType::class, $message);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_message_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute(
+                'app_message_index',
+                [],
+                Response::HTTP_SEE_OTHER
+            );
         }
 
         return $this->render('message/edit.html.twig', [
@@ -69,13 +115,33 @@ final class MessageController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_message_delete', methods: ['POST'])]
-    public function delete(Request $request, Message $message, EntityManagerInterface $entityManager): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$message->getId(), $request->getPayload()->getString('_token'))) {
+    public function delete(
+        Request $request,
+        Message $message,
+        EntityManagerInterface $entityManager
+    ): Response {
+
+        /** @var User $user */
+        $user = $this->getUser();
+
+        if ($message->getSender() !== $user) {
+            throw $this->createAccessDeniedException();
+        }
+
+        if (
+            $this->isCsrfTokenValid(
+                'delete' . $message->getId(),
+                $request->getPayload()->getString('_token')
+            )
+        ) {
             $entityManager->remove($message);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('app_message_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute(
+            'app_message_index',
+            [],
+            Response::HTTP_SEE_OTHER
+        );
     }
 }
